==> 03_attempt/2_recipes.R <==
# Regression Problem ----
# Setup preprocessing/recipes/feature engineering

# Load package(s) ----
library(tidyverse)
library(tidymodels)
library(here)
library(recipes)

# handle common conflicts
tidymodels_prefer()

# loading training data -------
load(here("03_attempt/data/training_data.rda"))

# Recipe 1
#recipe_1 <- recipe(log_price ~ ., data = training_data) |>
#  step_rm(id, host_has_profile_pic, property_type, host_has_profile_pic,
#          host_identity_verified, has_availability, instant_bookable) |>
#  step_impute_mean(all_numeric_predictors()) |>
#  step_impute_mode(all_nominal_predictors()) |>
#  step_dummy(all_nominal_predictors(), one_hot = TRUE) |>
#  step_nzv(all_predictors()) |>
#  step_normalize(all_numeric_predictors())

recipe_1 <- recipe(log_price ~ ., data = training_data) |>
  step_mutate(host_listings_ratio = host_listings_count / host_total_listings_count,
              host_has_profile_pic = factor(host_has_profile_pic, levels = c(FALSE, TRUE)),
              host_identity_verified = factor(host_identity_verified, levels = c(FALSE, TRUE)),
              has_availability = factor(has_availability, levels = c(FALSE, TRUE)), 
              instant_bookable = factor(instant_bookable, levels = c(FALSE, TRUE))) |>
  step_rm(host_listings_count, host_total_listings_count,
          id, host_verifications, host_response_time, first_review_year, last_review_year, 
          has_availability, instant_bookable, reviews_per_month, neighbourhood_cleansed)|>
  step_impute_mean(all_numeric_predictors()) |>
  step_impute_mode(all_nominal_predictors()) |>
  step_nzv(all_numeric_predictors()) |>
  step_dummy(all_nominal_predictors(), one_hot = TRUE) |>
  step_normalize(all_numeric_predictors())

# Recipe 2
recipe_2 <- recipe(log_price ~ ., data = training_data) |>
  step_rm(id, host_has_profile_pic, property_type, host_has_profile_pic,
          host_identity_verified, has_availability, instant_bookable,
          first_review_year, last_review_year, host_verifications, host_response_time,
          host_response_rate, reviews_per_month, neighbourhood_cleansed) |>
  step_impute_mean(all_numeric_predictors()) |>
  step_impute_mode(all_nominal_predictors()) |>
  step_dummy(all_nominal_predictors(), one_hot = TRUE) |>
  step_nzv(all_numeric_predictors()) |>
  step_normalize(all_numeric_predictors())

recipe_2 |>
  prep() |>
  bake(new_data = NULL) |>
  glimpse()

save(recipe_1, file = here("03_attempt/recipes/recipe_1.rda"))
save(recipe_2, file = here("03_attempt/recipes/recipe_2.rda"))
