==> 03_attempt/4_model_comparison_2.R <==
# Model analysis and comparison

# Load package(s) ----
library(tidyverse)
library(tidymodels)
library(tictoc)
library(here)
# Handle conflicts
tidymodels_prefer()

# load tune models ----
load(here("03_attempt/results/knn_tune_2.rda"))
load(here("03_attempt/results/bt_tune_2.rda"))
load(here("03_attempt/results/rf_tune_2.rda"))

# Select the best hyperparameters for random forest
# For rf model
rf_best <- select_best(rf_tune_1, metric = "mae")

# For bt
bt_best <- select_best(bt_tune_1, metric = "mae")

# For kknn model
knn_best <- select_best(knn_tune_1, metric = "mae")

# Creating table
# all models results
model_results <- as_workflow_set(
  knn_2 = knn_tune_2,
  rf_2 = rf_tune_2,
  bt_2 = bt_tune_2
)

models_table_2 <- model_results |>
  collect_metrics() |>
  filter(.metric == "mae") |>
  slice_min(mean, by = wflow_id) |>
  select(wflow_id, mean, std_err) |>
  arrange(mean)

model_results |>
  autoplot(metric = "mae")

model_results |>
  autoplot(metric = "mae", select_best = TRUE)
